Always use explicit integer dtypes in tests.
Possibly prevents problems caused by the 'int' dtype being mapped differently
depending on the platform.

+++ tests/test_backend_hdf5.py
@@ -15,7 +15,7 @@ def hdf5file(tmpdir):
 
 
 @pytest.mark.parametrize('schema_node,valid_data', (
-    (schema.Array(dtype='int'), np.array([23, 42])),
+    (schema.Array(dtype='int32'), np.array([23, 42], dtype='int32')),
     (schema.Bool(), True),
     (schema.Date(), datetime.date.today()),
     (schema.DateTime(), datetime.datetime.now()),
